split out controller tests into units of the individual controllers
